ui: extract TodoTask into component
